# OUT/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "PGSIMS API",
    "version": ""
  },
  "paths": {
...
    "/api/departments/": {
      "get": {
        "operationId": "listDepartments",
        "description": "",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "A page number within the paginated result set.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "A search term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "required": false,
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "results"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 123
                    },
                    "next": {
                      "type": "string",
                      "nullable": true,
                      "format": "uri",
                      "example": "http://api.example.org/accounts/?page=4"
                    },
                    "previous": {
                      "type": "string",
                      "nullable": true,
                      "format": "uri",
                      "example": "http://api.example.org/accounts/?page=2"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Department"
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "api"
        ]
      },
      "post": {
        "operationId": "createDepartment",
        "description": "",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/supervision-links/": {
      "get": {
        "operationId": "listSupervisorResidentLinks",
        "description": "",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "A page number within the paginated result set.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "A search term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "required": false,
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "results"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 123
                    },
                    "next": {
                      "type": "string",
                      "nullable": true,
                      "format": "uri",
                      "example": "http://api.example.org/accounts/?page=4"
                    },
                    "previous": {
                      "type": "string",
                      "nullable": true,
                      "format": "uri",
                      "example": "http://api.example.org/accounts/?page=2"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SupervisorResidentLink"
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "api"
        ]
      },
      "post": {
        "operationId": "createSupervisorResidentLink",
        "description": "",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupervisorResidentLink"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/SupervisorResidentLink"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SupervisorResidentLink"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupervisorResidentLink"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/users/": {
      "get": {
        "operationId": "listUsers",
        "description": "",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "A page number within the paginated result set.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "A search term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "required": false,
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "results"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 123
                    },
                    "next": {
                      "type": "string",
                      "nullable": true,
                      "format": "uri",
                      "example": "http://api.example.org/accounts/?page=4"
                    },
                    "previous": {
                      "type": "string",
                      "nullable": true,
                      "format": "uri",
                      "example": "http://api.example.org/accounts/?page=2"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserManagement"
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "api"
        ]
      },
      "post": {
        "operationId": "createUser",
        "description": "",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserManagement"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UserManagement"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UserManagement"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserManagement"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "api"
        ]
      }
    },
...
    "/academics/api/departments/": {
      "get": {
        "operationId": "listDepartments",
        "description": "ViewSet for Department CRUD operations.",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "A page number within the paginated result set.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "A search term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "required": false,
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count",
                    "results"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 123
                    },
                    "next": {
                      "type": "string",
                      "nullable": true,
                      "format": "uri",
                      "example": "http://api.example.org/accounts/?page=4"
                    },
                    "previous": {
                      "type": "string",
                      "nullable": true,
                      "format": "uri",
                      "example": "http://api.example.org/accounts/?page=2"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Department"
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "academics"
        ]
      },
      "post": {
        "operationId": "createDepartment",
        "description": "ViewSet for Department CRUD operations.",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "academics"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "Department name",
            "maxLength": 200
          },
          "code": {
            "type": "string",
            "description": "Department code",
            "maxLength": 20
          },
          "description": {
            "type": "string",
            "description": "Department description"
          },
          "head": {
            "type": "integer",
            "nullable": true,
            "description": "Department head"
          },
          "head_name": {
            "type": "string",
            "readOnly": true
          },
          "active": {
            "type": "boolean",
            "description": "Is department active?"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "name",
          "code"
        ]
      },
...
      "SupervisorResidentLink": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "supervisor_user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "readOnly": true
              },
              "username": {
                "type": "string",
                "description": "Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.",
                "pattern": "^[\\w.@+-]+\\z",
                "maxLength": 150
              },
              "full_name": {
                "type": "string",
                "readOnly": true
              },
              "role": {
                "enum": [
                  "admin",
                  "supervisor",
                  "pg",
                  "resident",
                  "faculty",
                  "utrmc_user",
                  "utrmc_admin"
                ],
                "type": "string",
                "description": "User role determines access permissions in SIMS"
              },
              "is_active": {
                "type": "boolean",
                "description": "Designates whether this user should be treated as active. Unselect this instead of deleting accounts."
              }
            },
            "required": [
              "username",
              "role"
            ],
            "readOnly": true
          },
          "resident_user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "readOnly": true
              },
              "username": {
                "type": "string",
                "description": "Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.",
                "pattern": "^[\\w.@+-]+\\z",
                "maxLength": 150
              },
              "full_name": {
                "type": "string",
                "readOnly": true
              },
              "role": {
                "enum": [
                  "admin",
                  "supervisor",
                  "pg",
                  "resident",
                  "faculty",
                  "utrmc_user",
                  "utrmc_admin"
                ],
                "type": "string",
                "description": "User role determines access permissions in SIMS"
              },
              "is_active": {
                "type": "boolean",
                "description": "Designates whether this user should be treated as active. Unselect this instead of deleting accounts."
              }
            },
            "required": [
              "username",
              "role"
            ],
            "readOnly": true
          },
          "department": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "readOnly": true
              },
              "name": {
                "type": "string",
                "description": "Department name",
                "maxLength": 200
              },
              "code": {
                "type": "string",
                "description": "Department code",
                "maxLength": 20
              },
              "active": {
                "type": "boolean",
                "description": "Is department active?"
              }
            },
            "required": [
              "name",
              "code"
            ],
            "readOnly": true
          },
          "supervisor_user_id": {
            "type": "integer",
            "writeOnly": true
          },
          "resident_user_id": {
            "type": "integer",
            "writeOnly": true
          },
          "department_id": {
            "type": "integer",
            "writeOnly": true,
            "nullable": true
          },
          "start_date": {
            "type": "string",
            "format": "date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "active": {
            "type": "boolean",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "supervisor_user_id",
          "resident_user_id",
          "start_date"
        ]
      },
...
      "UserManagement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "username": {
            "type": "string",
            "description": "Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.",
            "pattern": "^[\\w.@+-]+\\z",
            "maxLength": 150
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "minLength": 8
          },
          "first_name": {
            "type": "string",
            "maxLength": 150
          },
          "last_name": {
            "type": "string",
            "maxLength": 150
          },
          "full_name": {
            "type": "string",
            "readOnly": true
          },
          "role": {
            "enum": [
              "admin",
              "supervisor",
              "pg",
              "resident",
              "faculty",
              "utrmc_user",
              "utrmc_admin"
            ],
            "type": "string",
            "description": "User role determines access permissions in SIMS"
          },
          "specialty": {
            "enum": [
              "medicine",
              "surgery",
              "pediatrics",
              "gynecology",
              "orthopedics",
              "cardiology",
              "neurology",
              "urology",
              "psychiatry",
              "dermatology",
              "radiology",
              "anesthesia",
              "pathology",
              "microbiology",
              "pharmacology",
              "community_medicine",
              "forensic_medicine",
              "other"
            ],
            "type": "string",
            "nullable": true,
            "description": "Medical specialty (required for PGs and Supervisors)"
          },
          "year": {
            "enum": [
              "1",
              "2",
              "3",
              "4"
            ],
            "type": "string",
            "nullable": true,
            "description": "Training year (required for PGs)"
          },
          "is_active": {
            "type": "boolean",
            "description": "Designates whether this user should be treated as active. Unselect this instead of deleting accounts."
          },
          "supervisor": {
            "type": "integer",
            "nullable": true,
            "description": "Assigned supervisor (required for PGs)"
          },
          "home_department": {
            "type": "integer",
            "nullable": true,
            "description": "Primary home department for PG users until graduation"
          },
          "home_hospital": {
            "type": "integer",
            "nullable": true,
            "description": "Primary home hospital for PG users until graduation"
          },
          "registration_number": {
            "type": "string",
            "nullable": true,
            "description": "Medical council registration number",
            "maxLength": 50
          },
          "phone_number": {
            "type": "string",
            "nullable": true,
            "description": "Contact phone number",
            "maxLength": 15
          },
          "resident_profile": {
            "type": "object",
            "writeOnly": true
          },
          "staff_profile": {
            "type": "object",
            "writeOnly": true
          },
          "departments": {
            "type": "string",
            "readOnly": true
          },
          "date_joined": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "username",
          "role"
        ]
      },
...
    }
  }
}